Manifests: remove adapter-postgres dependency
