+++ packages/adapter-postgres/src/adapter/pg-connection-adapter.ts
@@ -1,11 +1,14 @@
 import { z } from "zod";
 
 import type { IConnectionAdapter } from "@rejot-dev/contract/adapter";
+import type {
+  PostgresConnectionSchema,
+  PostgresDataStoreSchema,
+} from "@rejot-dev/contract/manifest";
 import type { IConnection } from "@rejot-dev/contract/sync";
 
 import { PostgresEventStore } from "../event-store/postgres-event-store.ts";
 import { DEFAULT_PUBLICATION_NAME, DEFAULT_SLOT_NAME } from "../postgres-consts.ts";
-import type { PostgresConnectionSchema, PostgresDataStoreSchema } from "../postgres-schemas.ts";
 import { PostgresSink } from "../postgres-sink.ts";
 import { PostgresSource } from "../postgres-source.ts";
 import { PostgresClient } from "../util/postgres-client.ts";
